# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Tixist Standalone API
  version: 0.1.0
  description: |
    REST contract for the standalone event-management API extracted from the
    Next.js/tRPC application. Each tRPC procedure has been mapped to a
    resource-oriented REST endpoint; non-CRUD actions use sub-paths
    (e.g. `POST /events/{id}/archive`). This spec is the contract for the
    NestJS implementation.

    Authorization model:
      - Bearer JWT identifies the user.
      - Event-scoped routes require an ACTIVE TeamMember (owner or collaborator).
      - Module-gated routes require the relevant module in the collaborator's
        permissions (owners bypass). Modules: OVERVIEW, ATTENDEES, TICKETS,
        SCHEDULE, SPEAKERS, CFP, COMMUNICATIONS, CHECKIN. SETTINGS is owner-only.
      - Routes tagged `public` need no auth (event discovery, self-registration,
        public CFP, ticket lookup).
servers:
  - url: https://api.tixist.example.com/v1
    description: Production
  - url: http://localhost:4000/v1
    description: Local
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 6 }
                name: { type: string }
      responses:
        '201':
          {
            description: Created,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/User' } },
              },
          }
        '400': { $ref: '#/components/responses/BadRequest' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate with email + password, returns JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: { type: string }
                  refreshToken: { type: string }
                  user: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange a refresh token for a new token pair (rotation)
      description: 'Send the refresh token as `Authorization: Bearer <refreshToken>`.'
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: { type: string }
                  refreshToken: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke the current refresh token (access-protected)
      responses:
        '204': { description: Logged out }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ───────────────────────── Users (me) ─────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cronSecret:
      type: apiKey
      in: header
      name: Authorization
      description: 'Bearer {CRON_SECRET} for scheduler-triggered endpoints'
...
  responses:
    BadRequest:
      description: Validation error
      content:
        {
          application/problem+json:
            { schema: { $ref: '#/components/schemas/Problem' } },
        }
    Unauthorized:
      description: Missing/invalid credentials
      content:
        {
          application/problem+json:
            { schema: { $ref: '#/components/schemas/Problem' } },
        }
...
  schemas:
    Problem:
      type: object
      description: RFC 7807 problem details
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        errors: { type: object, additionalProperties: true }

    # ── Enums ──
...
    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string, nullable: true }
        email: { type: string, format: email, nullable: true }
        image: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
...